Don't play the video automatically when next question is shown
This may not be the optimal UI, but on Android Chrome video-playing event gets
fired but the actual video doesn't start playing because the browser doesn't
allow videos to play unless initiated by user actions.  This would cause other
event triggers to be off-sync and introduces bugs. To work around this, the
user is forced to tap on the video to start playing it, even after clicking the
"Next question" button.